Moved search form to own file
