+++ src/index.js
@@ -5,36 +5,8 @@ import axios from 'axios';
 import './index.css';
 import 'foundation-sites/dist/css/foundation.css';
 import Categories from './category.jsx';
+import SearchForm from './searchform.jsx';
 import SearchResults from './searchresult.jsx';
-import { Button, Sizes, Label, Row } from 'react-foundation';
-class SearchForm extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {value: '', results: []};
-    this.handleChange = this.handleChange.bind(this);
-    this.handleSubmit = this.handleSubmit.bind(this);
-  }
-  handleChange(event) {
-    this.setState({value: event.target.value});
-  }
-  handleSubmit(event) {
-    this.props.handler(this.state.value);
-    event.preventDefault();
-  }
-  render() {
-    return(
-      <Row>
-        <form onSubmit={this.handleSubmit}>
-          <Label>
-            keywords:
-          </Label>
-          <input type="text" value={this.state.value} onChange={this.handleChange} />
-          <Button isExpanded size={Sizes.LARGE} onSubmit={this.handleSubmit}>search</Button>
-        </form>
-      </Row>
-    );
-  }
-}
 
 class App extends React.Component {
   constructor(props) {
